dashboard moved into dashboard js file
